backend/requirements.txt
alembic==1.14.0
aniso8601==9.0.1
bcrypt==4.2.0
blinker==1.9.0
click==8.1.7
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
flask-serialize==2.2.0
Flask-SQLAlchemy==3.1.1
flask_serializer==0.0.5.1
greenlet==3.1.1
gunicorn==23.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
packaging==24.2
permissive-dict==1.0.4
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.3
pytest-flask==1.3.0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
tomli==2.0.2
typing_extensions==4.12.2
Werkzeug==3.1.3
